% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cheeseQualityData}
\alias{cheeseQualityData}
\title{Experts' opinions about cheese quality}
\format{
A list of 120 trapezoidal fuzzy numbers (as defined in \code{FuzzyNumbers} package) for three experts (40 answers of the
first expert, 38 of the second, and 42 of the third one).
Each observation in the list has its label \code{Exp.y.z}, where \code{y} is the number of the expert, and \code{z} is the
number of his answer.
}
\source{
Ramos-Guajardo, A.B., Blanco-Fernandez, A., Gonzalez-Rodriguez, G. (2019)
Applying Statistical Methods with Imprecise Data to Quality Control in Cheese Manufacturing
<https://doi.org/10.1007/978-3-030-03201-2_8>
}
\usage{
cheeseQualityData
}
\description{
A set of fuzzy epistemic data containing the corrected version of the three experts' opinions about the Gamonedo cheese quality data.
}
\keyword{datasets}
